Update rearm to give price based on weapon.
Added a new function to get prices for each weapon in the vehicle. Default = ["ItemGoldBar10oz",2] unless specified in _rearm_costs[].

I didn't think it was fair to allow a vehicle to have rockets rearmed for the same price as a machine gun. This function now allows unique prices based on weapon.